Do not os.Exit(1)
This makes sense if you're running a script.
If you're running a server, however, this means an HTTP parsing error
would attempt to shut down your application.

Instead, return the error and let the consumer decide what to do with
it.

This is backwards incompatible, and probably deserves a version bump.